Alpha Base 0.03
Made a Quick easy test place in my Scene.
Added a Small movement script, that is binded to the player.
Still working on setting up UI for HP, for now.
Will make a test script for Fire (stand in the object and take HP lose) later when I got the UI working.